+++ R/polar_freqMap.R
@@ -149,7 +149,7 @@ freqMap <- function(data,
   }
 
   # cut data
-  data <- quick_cutdata(data = data, type = type)
+  data <- openair::cutData(x = data, type = type %||% "default", ...)
 
   # deal with popups
   if (length(popup) > 1) {