+++ R/polar_polarMap.R
@@ -370,7 +370,7 @@ polarMap <- function(data,
   }
 
   # cut data
-  data <- quick_cutdata(data = data, type = type)
+  data <- openair::cutData(x = data, type = type %||% "default", ...)
 
   # deal with upper
   if (any(upper == "fixed")) {